Fix BodyDetails () function in Cottle not reporting Material correctly. Also fix rings not being reported correctly. Issue reported by gdt in EDCD/EDDI discord.
